+++ packages/opentelemetry-sdk-trace-node/test/NodeTracerProvider.test.ts
@@ -236,7 +236,7 @@ describe('NodeTracerProvider', () => {
     });
 
     it('should allow propagators as per the specification', () => {
-      (process.env as any).OTEL_PROPAGATORS = 'b3,b3multi,jaeger';
+      (process.env as any).OTEL_PROPAGATORS = 'b3,b3multi,jaeger,xray';
 
       const provider = new NodeTracerProvider();
       provider.register();
@@ -249,6 +249,7 @@ describe('NodeTracerProvider', () => {
         'x-b3-sampled',
         'x-b3-parentspanid',
         'uber-trace-id',
+        'x-amzn-trace-id',
       ]);
     });
   });